pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trade-system-modules"
version = "0.1.0"
description = "Shared modules for trade systems: config, data adapters, storage, execution"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "Hedge Fund Engineering", email = "engineering@hedgefund.com"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.0",
    "pandas>=2.0",
    "pyarrow>=12.0",
    "google-cloud-storage>=2.10",
    "ib-insync>=0.9.70",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
    "mypy>=1.5",
    "types-requests>=2.31",
]

[project.urls]
Homepage = "https://github.com/org/trade_system_modules"
Repository = "https://github.com/org/trade_system_modules.git"

[tool.setuptools.packages.find]
where = ["src"]